requirements.txt
ticktick-py=2.0.3